+++ fairscale/nn/data_parallel/fully_sharded_data_parallel.py
@@ -653,17 +653,21 @@ class FullyShardedDataParallel(nn.Module):
         if torch.cuda.is_available():
             torch.cuda.synchronize()
         self._lazy_init()
-        if self.mixed_precision:
-            # Buffers dtype stays consistent with parameters.
-            self._cast_buffers(dtype=torch.float32)
+
+        def maybe_cast_buffers(dtype: Optional[torch.dtype] = None) -> None:
+            if self.mixed_precision:
+                self._cast_buffers(dtype=dtype)
 
         if self._return_full_state_dict:
             if self.training_state != TrainingState.SUMMON_FULL_PARAMS:
                 with self.summon_full_params(recurse=False, volatile=True):
+                    maybe_cast_buffers(torch.float32)
                     state_dict = super().state_dict(*args, **kwargs)
             else:
+                maybe_cast_buffers(torch.float32)
                 state_dict = super().state_dict(*args, **kwargs)
         else:
+            maybe_cast_buffers(torch.float32)
             if self.flatten_parameters:
                 assert isinstance(self.module, FlattenParamsWrapper)
                 state_dict = self.module.flat_state_dict(*args, **kwargs)
@@ -674,9 +678,8 @@ class FullyShardedDataParallel(nn.Module):
             for k in state_dict.keys():
                 state_dict[k] = state_dict[k].cpu()
 
-        if self.mixed_precision:
-            # In case we are in mixed precision, restore buffers back to buffer_dtype.
-            self._cast_buffers()
+        # In case we are in mixed precision, restore buffers back to buffer_dtype.
+        maybe_cast_buffers()
         return state_dict
 
     @typing.overload
@@ -830,7 +833,7 @@ class FullyShardedDataParallel(nn.Module):
                     # latter may contain padding.
                     assert len(self.params) == 1
                     assert isinstance(self.module, FlattenParamsWrapper)
-                    stack.enter_context(self.module.unflatten_params(recurse=False, flat_param=self.params[0]))
+                    stack.enter_context(self.module.unflatten_params(flat_param=self.params[0]))
                 try:
                     yield
                 finally:
@@ -860,7 +863,7 @@ class FullyShardedDataParallel(nn.Module):
 
     def _lazy_init(self) -> None:
         """Initialization steps that should happen lazily, typically right
-           before the first forward pass.
+        before the first forward pass.
         """
         # Initialize param attributes lazily, in case the param's dtype or
         # device changes after __init__.
@@ -1531,7 +1534,7 @@ class FullyShardedDataParallel(nn.Module):
             # There are as many sharded parameters as there parameters in the
             # consolidated model, so we only need to export how to reshape the
             # parameters to their orginal shape and take care of the padding
-            if not hasattr(m, "_param_numels"):
+            if not m.flatten_parameters:
                 params_metadata.append(
                     {
                         "fsdp_path": _clean_path(path),
@@ -1560,8 +1563,11 @@ class FullyShardedDataParallel(nn.Module):
                         "is_flat": True,
                         "num_padded": m.numel_padded_per_param,
                         "param_names": param_names,
-                        "param_shapes": m._param_shapes,
-                        "param_numels": m._param_numels,
+                        # TODO (Min): we don't want to access the private _param_shapes and
+                        # _param_numels here. We want to dump metadata from FPW when there are
+                        # multiple groups of params.
+                        "param_shapes": m._fsdp_wrapped_module.flat_param._param_shapes,  # type: ignore
+                        "param_numels": m._fsdp_wrapped_module.flat_param._param_numels,  # type: ignore
                         "no_broadcast_optim_state": m.no_broadcast_optim_state,
                     }
                 )
@@ -1586,7 +1592,7 @@ class FullyShardedDataParallel(nn.Module):
         If this behavior is not correct for your module (for instance if buffers
         needs to be reduced instead), you can disable it with `with_module_buffers=False`.
 
-        This method is very useful to re-assemble checkpoints of shards without
+        This method is used to re-assemble checkpoints of shards without
         having to instantiate FSDP wrappers with the world size originally used
         to save the shards.
         """