+++ tests/nn/checkpoint/test_checkpoint_activations.py
@@ -52,7 +52,7 @@ def get_loss_and_gnorm(model, input):
 class BasicModel(nn.Module):
     """Basic model with a single FFN being checkpointed.
 
-       Used for extensive checkings: equivalency with non-checkpoint, torch-checkpoint, etc.
+    Used for extensive checkings: equivalency with non-checkpoint, torch-checkpoint, etc.
     """
 
     def __init__(self, use_pytorch_checkpoint=False, use_fairscale_checkpoint=False, **kwargs):
@@ -267,3 +267,38 @@ def test_deprecated_path():
     # Check if direct import works as before.
     ffn = nn.Sequential(nn.Linear(32, 128), nn.Dropout(p=0.5), nn.Linear(128, 32),)
     ffn = deprecated_checkpoint_wrapper(ffn, {})
+
+
+@skip_if_no_cuda
+def test_list_input():
+    """ Test checkpointing with input argument type being a list.
+
+    Note: Testing shows that PyTorch's torch.utils.checkpoint function does not pass this test.
+    """
+    count = 0
+
+    class Model(nn.Module):
+        def __init__(self):
+            super().__init__()
+            self.conv = nn.Linear(2, 2)
+
+        def forward(self, x):
+            nonlocal count
+            count += 1
+            y = []
+            for i in x:
+                y.append(self.conv(i))
+            return y
+
+    model = nn.Sequential(checkpoint_wrapper(Model()), Model()).cuda()
+    in_data1 = torch.rand(4, 2).cuda()
+    in_data2 = torch.rand(4, 2).cuda()
+
+    # Forward. Count should be 2 for 2 modules.
+    out = model([in_data1, in_data2])
+    loss = sum(x.sum() for x in out)
+    assert count == 2, f"Incorrect count {count}"
+
+    # Backward. Adds 1 more forward call due to checkpoint.
+    loss.backward()
+    assert count == 3, f"Incorrect count {count}"